Changed file structure to meet dart standards
