+++ lib/src/manager.dart
