+++ lib/src/sprinkle.dart
@@ -1,0 +1,17 @@
+import 'package:flutter/material.dart';
+
+import 'package:sprinkle/sprinkle.dart';
+
+class Sprinkle extends StatelessWidget {
+  final Widget child;
+  final Supervisor supervisor;
+
+  Sprinkle({this.child, this.supervisor});
+
+  Widget build(BuildContext context) {
+    return Provider(
+      data: supervisor,
+      child: child,
+    );
+  }
+}
