+++ pubspec.yaml
@@ -1,5 +1,5 @@
 name: sprinkle
-description: A convenient, stream-based Flutter architecture
+description: A convenient and easy to use, stream-based Flutter architecture
 version: 0.0.5
 homepage: https://github.com/zaiste/sprinkle
